feat: update word checking method
